+++ src/qq_lib/batch/interface/interface.py
@@ -759,6 +759,36 @@ class BatchInterface[
         jobs.sort(key=lambda job: job.getId())
 
     @classmethod
+    def jobsPresenterColumnsToShow(cls) -> set[str]:
+        """
+        Get a set of columns that should be shown in the output of JobsPresenter (`qq jobs`)
+        for this batch system.
+
+        In the default implementation, all columns are shown.
+
+        Note that the 'Exit' column is not shown when printing queued and running jobs,
+        even if you specify it here.
+
+        Args:
+            set[str]: Set of column titles that should be shown.
+        """
+        return {
+            "S",
+            "Job ID",
+            "User",
+            "Job Name",
+            "Queue",
+            "NCPUs",
+            "NGPUs",
+            "NNodes",
+            "Times",
+            "Node",
+            "%CPU",
+            "%Mem",
+            "Exit",
+        }
+
+    @classmethod
     def _translateSSHCommand(cls, host: str, directory: Path) -> list[str]:
         """
         Construct the SSH command to navigate to a remote directory.